feat: flaskのディレクトリ構成を修正
